\name{Bidart-package}
\alias{Bidart-package}
\alias{Bidart}
\docType{package}
\title{
\packageTitle{Bidart}
}
\description{
\packageDescription{Bidart}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Bidart}
\packageIndices{Bidart}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Bidart}

Maintainer: \packageMaintainer{Bidart}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
