% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{compare_PP_NNGP}
\alias{compare_PP_NNGP}
\title{Allows to see if a Predictive Process has enough knots.  
Plots two samples, one from a Predictive Process, and one from the Nearest Neighbor Gaussian Process the PP is obtained from. 
If there are not enough knots, the PP should be over-smoothed with respect to the NNGP.}
\usage{
compare_PP_NNGP(PP)
}
\arguments{
\item{PP}{a Predictive Process object}
}
\description{
Allows to see if a Predictive Process has enough knots.  
Plots two samples, one from a Predictive Process, and one from the Nearest Neighbor Gaussian Process the PP is obtained from. 
If there are not enough knots, the PP should be over-smoothed with respect to the NNGP.
}
