% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Useful_stuff.R
\name{compute_sparse_chol}
\alias{compute_sparse_chol}
\title{Computes a Vecchia sparse Cholesky factor}
\usage{
compute_sparse_chol(
  range_beta,
  NNarray,
  locs,
  range_X = NULL,
  PP = NULL,
  use_PP = F,
  compute_derivative = T,
  nu = 1.5,
  anisotropic = F,
  sphere = F,
  num_threads = 1,
  locs_idx = NULL
)
}
\arguments{
\item{range_beta}{parameter for the range. If a stationary covariance is used, the parameter is passed to the exponential. Else,}

\item{NNarray}{Vecchia parents array provided by GpGp::find_ordered_nn}

\item{locs}{matrix of spatial sites}

\item{range_X}{covariates for range}

\item{PP}{predictive process obtained through get_PP}

\item{use_PP}{should the PP be used ?}

\item{compute_derivative}{logical, indicates if derivatives of Vecchia factors are to be computed}

\item{nu}{Matern smoothness}

\item{locs_idx}{match between PP basis function and locs}
}
\description{
Computes a Vecchia sparse Cholesky factor
}
