% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_nngp_predict_nonstationary.R
\name{DIC}
\alias{DIC}
\title{Deviance Information Criterion}
\usage{
DIC(mcmc_nngp_list, burn_in = 0.5)
}
\arguments{
\item{mcmc_nngp_list}{A mcmc_nngp_list object generated using mcmc_nngp_initialize and ran using mcmc_nngp_run}

\item{burn_in}{MCMC burn-in}
}
\description{
Deviance Information Criterion
}
