% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{ESS}
\alias{ESS}
\title{Prints the Effective Sample Size of a MCMC run.
The proportion of the burn-in is adjustable.}
\usage{
ESS(mcmc_nngp_list, burn_in = 0.5)
}
\arguments{
\item{mcmc_nngp_list}{a mcmc_nngp_list created and run by the package}

\item{burn_in}{between 0.01 and .99, the proportion of samples discarded for the burn-in}
}
\description{
Prints the Effective Sample Size of a MCMC run.
The proportion of the burn-in is adjustable.
}
