% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_nngp_predict_nonstationary.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Estimates the parameters from mcmc_nngp_list}
\usage{
estimate_parameters(
  mcmc_nngp_list,
  burn_in = 0.5,
  get_samples = F,
  lib.loc = NULL
)
}
\arguments{
\item{mcmc_nngp_list}{A mcmc_nngp_list object generated using mcmc_nngp_initialize and ran using mcmc_nngp_run}

\item{burn_in}{MCMC burn-in}

\item{get_samples}{return Samples of the latent field}

\item{lib.loc}{Location of libraries if installed in manually specified folder}
}
\description{
Estimates the parameters from mcmc_nngp_list
}
