% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_nngp_predict_nonstationary.R
\name{log_score_Gaussian}
\alias{log_score_Gaussian}
\title{Log-density of observations with respect to the model.
Gives a training score when the observations are used in the model.
Gives a validation score when the observations are not used in the model.}
\usage{
log_score_Gaussian(
  observed_field,
  latent_field_samples,
  log_noise_samples,
  fixed_effects_samples
)
}
\arguments{
\item{latent_field_samples}{Samples of the latent field, either given by predict_latent_field or estimate_parameters}

\item{log_noise_samples}{Samples of the log noise variance, given by predict_noise}

\item{fixed_effects_samples}{Samples of the fixed effects, given by predict_fixed_effects}

\item{mcmc_nngp_list}{A mcmc_nngp_list object generated using mcmc_nngp_initialize and ran using mcmc_nngp_run}

\item{burn_in}{MCMC burn-in}
}
\description{
Log-density of observations with respect to the model.
Gives a training score when the observations are used in the model.
Gives a validation score when the observations are not used in the model.
}
