% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_ellipses}
\alias{plot_ellipses}
\title{Plots range ellipses for nonstationary covariance functions.}
\usage{
plot_ellipses(locs, log_range, shrink = 0.1, main = "ellipses", add = F)
}
\arguments{
\item{locs}{ellipses centers}

\item{log_range}{log_range at the ellipse centers, can hae 1 or 3 columns}

\item{shrink}{shrinks or inflates the ellipses}

\item{main}{main title}

\item{add}{ad on existing plot ? cancels main}
}
\description{
Plots range ellipses for nonstationary covariance functions.
}
