% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_pointillist_painting}
\alias{plot_pointillist_painting}
\title{Plots a spatial variable like a pointillist painting
using R base's points. Stupid, but handy.}
\usage{
plot_pointillist_painting(locs, field, cex = 1, main = NULL)
}
\arguments{
\item{locs}{spatial locations}

\item{field}{interest variable}

\item{cex}{shrinks or inflates the points}

\item{main}{main title}
}
\description{
Plots a spatial variable like a pointillist painting
using R base's points. Stupid, but handy.
}
