% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_nngp_predict_nonstationary.R
\name{predict_fixed_effects}
\alias{predict_fixed_effects}
\title{Predicts the fixed effects, whatever the locations}
\usage{
predict_fixed_effects(
  mcmc_nngp_list,
  X_pred = NULL,
  burn_in = 0.5,
  lib.loc = NULL
)
}
\arguments{
\item{mcmc_nngp_list}{A mcmc_nngp_list object generated using mcmc_nngp_initialize and ran using mcmc_nngp_run}

\item{X_pred}{New covariates}

\item{burn_in}{MCMC burn-in}

\item{lib.loc}{Location of libraries if installed in manually specified folder}
}
\description{
Predicts the fixed effects, whatever the locations
}
