% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_nngp_predict_nonstationary.R
\name{predict_latent_field}
\alias{predict_latent_field}
\title{# Actually predicts more than the latent field}
\usage{
predict_latent_field(
  mcmc_nngp_list,
  predicted_locs,
  X_range_pred = NULL,
  X_scale_pred = NULL,
  burn_in = 0.5,
  n_cores = 1,
  predict_range = F,
  predict_scale = F
)
}
\description{
# Actually predicts more than the latent field
}
